added create method in the bulk_discounts_controller and changed a migration for the bulk_discounts column to be an integer rather than a float

+++ db/migrate/20230926163846_create_bulk_discounts.rb
@@ -2,7 +2,7 @@ class CreateBulkDiscounts < ActiveRecord::Migration[7.0]
   def change
     create_table :bulk_discounts do |t|
       t.integer :quantity_threshold
-      t.float :percentage_discount
+      t.integer :percentage_discount
       t.references :merchant, null: false, foreign_key: true
 
       t.timestamps
